# Fix for #21456 (Regressions in attribute allocations for non-generic attributes)