--- a/src/System.Private.CoreLib/src/System/Reflection/CustomAttribute.cs
+++ b/src/System.Private.CoreLib/src/System/Reflection/CustomAttribute.cs
@@ -1615,9 +1615,10 @@ namespace System.Reflection
 
             if (ctorHasParameters)
             {
+                // Resolve method ctor token found in decorated decoratedModule scope
+                // See https://github.com/dotnet/coreclr/issues/21456 for why we fast-path non-generics here (fewer allocations)
                 if (attributeType.IsGenericType)
                 {
-                    // Resolve method ctor token found in decorated decoratedModule scope
                     ctor = decoratedModule.ResolveMethod(caRecord.tkCtor, attributeType.GenericTypeArguments, null).MethodHandle.GetMethodInfo();
                 }
                 else